Fix bug where board was continuously redrawing
The board was continuously redrawing because because getCursorRowCol
sometimes returned -1 for one value and a real value for another. We
would incorrectly interpret this as a valid cursor change.

Switch to return a *Coordinate instead so we can use nil as the invalid
value.

Also, explicitly redraw the board when an algorithm makes a move. We
where relying on this bug to redraw the board for the algorithms.